feat: add BigDeciamal to count

+++ src/main/java/br/com/api_str_innovation/dto/delivery/DeliveryReportDTO.java
@@ -1,5 +1,6 @@
 package br.com.api_str_innovation.dto.delivery;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -8,13 +9,13 @@ public record DeliveryReportDTO(
         String clientName,
         List<ProductInfo> products,
         Double totalWeight,
-        Double totalPaid,
+        BigDecimal totalPaid,
         LocalDate createdDt
 ) {
     public record ProductInfo(
             String name,
             Integer quantity,
             Double weight,
-            Double price
+            BigDecimal price
     ) {}
 }
